Refactor: Migrate components and pages to new structure; remove unused components
- Deleted old components: Reading, AdminNoteCard, AdminNoteGrid, Footer, Header, Hero, ModalNoteForm, NotesGrid.
- Created new components and pages under the (home) and admin directories.
- Implemented new layout for home page with improved note fetching and rendering.
- Added API route for login functionality with cookie management.
- Introduced middleware for authentication checks on admin routes.
- Updated utility function to get base URL dynamically.
