+++ src/app/(home)/footer.tsx
@@ -1,0 +1,12 @@
+const Footer = () => {
+    return (
+        <footer className="bg-darkgreen-800 text-white py-8 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 justify-items-center px-4 bottom-0">
+            <p>© 2025 NotesHub. All rights reserved.</p>
+            <p>Rahul Das</p>
+            <p>BWU/BCA/23/406</p>
+            <p>Mobile: +919775150082</p>
+        </footer>
+    );
+};
+
+export default Footer;
