+++ src/app/layout.tsx
@@ -1,9 +1,18 @@
+import Footer from "@/app/(home)/footer";
+import Header from "@/app/(home)/header";
+import { Toaster } from "@/components/ui/sonner";
+import { Analytics } from "@vercel/analytics/next";
 import type { Metadata } from "next";
+import { Montserrat } from "next/font/google";
 import "./globals.css";
-import Header from "@/components/header";
-import Footer from "@/components/footer";
-import { Toaster } from "@/components/ui/sonner"
-import { Analytics } from "@vercel/analytics/next"
+
+const montserrat = Montserrat({
+    subsets: ["latin"],
+    variable: "--font-montserrat",
+    weight: ["300", "400", "500", "600", "700"],
+    display: "swap",
+});
+
 
 export const metadata: Metadata = {
     title: "NotesHub | Notes Sharing Platform",
@@ -17,7 +26,7 @@ export default function RootLayout({
 }>) {
     return (
         <html lang="en">
-            <body>
+            <body className={`${montserrat.variable} antialiased scroll-smooth`}>
                 <main className="min-h-screen flex flex-col bg-darkgreen-50">
                     <Header />
                     {children}
